% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatrix.R
\name{random_matrix}
\alias{random_matrix}
\title{Create a matrix}
\usage{
random_matrix(n_row = 2, n_col = 2, set_mean = 0, set_sd = 5)
}
\arguments{
\item{n_row}{customize the matrix size by row number}

\item{n_col}{customize the matrix size by column number}

\item{set_mean}{customize the mean while generating the random number}

\item{set_sd}{customize the standard deviation while generating the random number}
}
\value{
A matrix
}
\description{
Create a matrix
}
\examples{
x <- random_matrix(n_row = 3, n_col = 5, set_mean = 0, set_sd = 1)
}
